Move internals files into framework folder
* Add required config to pull from this folder
