+++ src/framework/internals/src/lambda.ts
@@ -3,7 +3,7 @@ import {
   APIGatewayProxyEventV2WithRequestContext,
   APIGatewayProxyHandlerV2
 } from "aws-lambda";
-import {ActionManifest, ActionPayload} from "../../../framework/service/types";
+import {ActionManifest, ActionPayload} from "./types";
 import {inspect} from "util";
 
 export class RequestError extends Error {
@@ -20,11 +20,11 @@ export const LambdaExtractPayload = (event: APIGatewayProxyEventV2WithRequestCon
   try {
     payload = Object.assign(payload, JSON.parse(
       event.isBase64Encoded ?
-        Buffer.from(event.body, 'base64').toString('ascii') :
-        event.body
+        Buffer.from(event.body ? event.body : '', 'base64').toString('ascii') :
+        event.body ?? ''
     ));
   } catch (e) {
-    if (event.body.length !== 0) {
+    if (event.body && event.body.length !== 0) {
       throw new RequestError('The request body is malformed.');
     }
   }
@@ -45,8 +45,8 @@ export const LambdaExtractPayload = (event: APIGatewayProxyEventV2WithRequestCon
 export const stringifyResponse = (input: unknown) => {
   if (!input) return;
 
-  const allKeys = [];
-  const seen = {};
+  const allKeys: any[] = [];
+  const seen: any = {};
 
   JSON.stringify(input, (key, value) => {
     if (!(key in seen)) {
@@ -74,38 +74,41 @@ export const LambdifyHandler = (Handler: ActionManifest['Handler']): APIGatewayP
       body: stringifyResponse(response),
     };
   } catch (e) {
-    console.log(inspect({
-      stack: e.stack,
-      code: e.code,
-      message: e.message,
-      error: e.constructor.name,
-    }));
-
-    switch (e.constructor) {
-      case RequestError:
-        return {
-          statusCode: 400,
-          isBase64Encoded: false,
-          body: JSON.stringify({
-            message: e.message,
-          }),
-        };
-      case ServerError:
-        return {
-          statusCode: 500,
-          isBase64Encoded: false,
-          body: JSON.stringify({
-            message: e.message,
-          }),
-        }
-      default:
-        return {
-          statusCode: 500,
-          isBase64Encoded: false,
-          body: JSON.stringify({
-            message: e.message,
-          }),
-        };
+    if (e instanceof Error) {
+      switch (e.constructor) {
+        case RequestError:
+          return {
+            statusCode: 400,
+            isBase64Encoded: false,
+            body: JSON.stringify({
+              message: e.message,
+            }),
+          };
+        case ServerError:
+          return {
+            statusCode: 500,
+            isBase64Encoded: false,
+            body: JSON.stringify({
+              message: e.message,
+            }),
+          }
+        default:
+          return {
+            statusCode: 500,
+            isBase64Encoded: false,
+            body: JSON.stringify({
+              message: e.message,
+            }),
+          };
+      }
     }
+
+    return {
+      statusCode: 500,
+      isBase64Encoded: false,
+      body: JSON.stringify({
+        message: e,
+      }),
+    };
   }
 };
