+++ src/framework/internals/tsconfig.json
@@ -1,14 +1,12 @@
 {
+  "extends": "@sindresorhus/tsconfig",
   "compilerOptions": {
     "target": "es2015",
     "module": "commonjs",
     "moduleResolution": "node",
     "esModuleInterop": true,
     "resolveJsonModule": true,
-    "jsx": "react",
-    "paths": {
-      "internals/*": ["../../framework/internals/src/*"]
-    }
+    "declaration": false
   },
   "exclude": ["**/*.test.ts"]
 }
