--- src/services/search/.babelrc
@@ -1,6 +1,0 @@
-{
-  "presets": [
-    ["@babel/preset-env", {"targets": {"node": "current"}}],
-    "@babel/preset-typescript"
-  ]
-}
