+++ src/services/search/actions/drop-index.ts
@@ -1,6 +1,6 @@
-import {ActionPayload, ActionResponse} from "../../../framework/service/types";
+import {ActionPayload, ActionResponse} from "internals/types";
 import {dropIndex} from "../lib/opensearch";
-import {LambdifyHandler, RequestError} from "../lib/lambda";
+import {LambdifyHandler, RequestError} from "internals/lambda";
 import {inspect} from "util";
 
 export const Name = 'DropIndex';
