+++ src/services/search/babel.config.js
@@ -1,0 +1,14 @@
+module.exports = {
+  plugins: [
+    ["module-resolver", {
+      root: ["./"],
+      alias: {
+        internals: "../../framework/internals/src/"
+      }
+    }]
+  ],
+  presets: [
+    ["@babel/preset-env", {targets: {node: "current"}}],
+    "@babel/preset-typescript"
+  ]
+}
