+++ src/services/search/lib/Environment.ts
@@ -1,0 +1,2 @@
+export const getVariable =
+  (name: string): string => process.env[name] || '';
