+++ src/services/search/package.json
@@ -17,9 +17,11 @@
     "@babel/core": "^7.16.5",
     "@babel/preset-env": "^7.16.5",
     "@babel/preset-typescript": "^7.16.5",
+    "@sindresorhus/tsconfig": "^2.0.0",
     "@types/aws-lambda": "^8.10.89",
     "@types/jest": "^27.4.0",
     "babel-jest": "^27.4.5",
+    "babel-plugin-module-resolver": "^4.1.0",
     "faker": "^5.5.3",
     "jest": "^27.4.5",
     "ts-loader": "^9.2.6",
